Добавляет функционал сортировки фильмов
